refactor new file dispatch/http_api_client.go
